Implementation of cache creation with Spring Cache Abstraction

+++ src/main/java/com/compass/application/services/ProductService.java
@@ -9,6 +9,8 @@ import com.compass.application.services.exceptions.ObjectNotFoundException;
 import com.compass.application.services.exceptions.ProductAlreadyExistsException;
 import com.compass.application.services.exceptions.ProductInSaleException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
@@ -23,28 +25,32 @@ public class ProductService {
     @Autowired
     private StockService stockService;
 
+    @Cacheable(value = "products", key = "#root.methodName")
     public List<Product> findAll() {
         return productRepository.findAll();
     }
 
+    @Cacheable(value = "products", key = "#id")
     public Product findById(Long id) {
         return productRepository.findById(id).orElseThrow(() -> new ObjectNotFoundException("Not found Product: " + id));
     }
 
+    @CacheEvict(value = "products", key = "'findAll'")
     public Product save(ProductDTO productDTO) {
         try {
             Product product = new Product(null, productDTO.name(), productDTO.price(), productDTO.enabled(), null);
             product = productRepository.save(product);
-            stockService.save(new Stock(null, product, productDTO.quantity() != null ? productDTO.quantity() : 0));
+            stockService.save(new Stock(product.getId(), product, productDTO.quantity() != null ? productDTO.quantity() : 0));
             return product;
-        } catch (DataIntegrityViolationException e){
+        } catch (DataIntegrityViolationException e) {
             throw new ProductAlreadyExistsException("Product Already Exists on stock");
         }
     }
 
+    @CacheEvict(value = "products", allEntries = true)
     public Product update(Long id, ProductDTO productDTO) {
         Product product = findById(id);
-        Stock stock = stockService.findById(productDTO.id());
+        Stock stock = stockService.findById(id);
         product.setName(productDTO.name());
         product.setPrice(productDTO.price());
         product.setEnabled(productDTO.enabled());
@@ -54,18 +60,27 @@ public class ProductService {
         return product;
     }
 
-    // if product is already enabled then update to disabled
+    @CacheEvict(value = "products", allEntries = true)
     public Product enableOrDisable(Long id, EnableProductDTO enableProductDTO) {
         Product product = findById(id);
-        product.setEnabled(true);
+        product.setEnabled(enableProductDTO.enabled());
         return productRepository.save(product);
     }
 
+    @CacheEvict(value = "products", allEntries = true)
     public void delete(Long id) {
+        if (!productRepository.existsById(id)) {
+            throw new ObjectNotFoundException("Not found Product: " + id);
+        }
+
         Product product = findById(id);
+        Stock stock = stockService.findById(id);
+        stockService.delete(stock.getId());
+
         try {
             productRepository.deleteById(id);
         } catch (DataIntegrityViolationException e) {
+            stockService.save(stock);
             throw new ProductInSaleException("Cannot delete Product with active sale");
         }
     }
